requirements.txt
langgraph
langchain
langchain-core
langchain-groq
requests
joblib
scikit-learn
python-dotenv
typing-extensions
streamlit